Move loading_screen
