+++ src/lib/loadingscreen/index.css
@@ -1,0 +1,32 @@
+h1 {
+	font-family: mojangles, minecraft, monospace;
+}
+.loader {
+	display: initial;
+}
+#loading-image {
+	height: 75%;
+	top: 50%;
+	left: 50%;
+	position: absolute;
+	transform: translate(-50%, -50%);
+	image-rendering: crisp-edges;
+	image-rendering: -webkit-crisp-edges;
+}
+
+#loading-background {
+	background-color: #60a490;
+	z-index: 100;
+	height: 100% !important;
+	width: 100%;
+	position: fixed;
+}
+
+#loading-text {
+	color: #29594b;
+	z-index: 200;
+	position: absolute;
+	top: 50%;
+	left: 50%;
+	transform: translate(-50%, 12rem);
+}
